first pass tidy and implementation for withdrawals
