+++ cpg_workflows/large_cohort/combiner.py
@@ -9,18 +9,110 @@ Inputs:
     - vds_paths: list[str] | None - The optional list of VDS paths in string format to combine
 """
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from hail.vds import new_combiner
 from hail.vds.combiner.variant_dataset_combiner import VariantDatasetCombiner
+from hailtop.batch.job import PythonJob
 
-from cpg_utils.config import config_retrieve
-from cpg_utils.hail_batch import init_batch
+from cpg_utils.config import config_retrieve, genome_build
+from cpg_utils.hail_batch import get_batch, init_batch
 from cpg_workflows.batch import override_jar_spec
-from cpg_workflows.utils import can_reuse, to_path
+from cpg_workflows.targets import Cohort, SequencingGroup
+from cpg_workflows.utils import can_reuse, slugify, to_path
+from metamist.graphql import gql, query
+
+if TYPE_CHECKING:
+    from graphql import DocumentNode
 
 
-def run(
+def _initalise_combiner_job(cohort: Cohort) -> PythonJob:
+    j: PythonJob = get_batch().new_python_job('Combiner', (cohort.get_job_attrs() or {}) | {'tool': 'hail query'})  # type: ignore[reportUnknownArgumentType]
+    j.image(config_retrieve(['workflow', 'driver_image']))
+    j.memory(config_retrieve(['combiner', 'memory']))
+    j.storage(config_retrieve(['combiner', 'storage']))
+
+    # set this job to be non-spot (i.e. non-preemptible)
+    # previous issues with preemptible VMs led to multiple simultaneous QOB groups processing the same data
+    j.spot(config_retrieve(['combiner', 'preemptible_vms'], False))
+    return j
+
+
+def get_vds_ids_output(vds_id: int) -> tuple[str, list[str]]:
+    get_vds_analysis_query: DocumentNode = gql(
+        """
+            query getVDSByAnalysisIds($vds_id: Int!) {
+                analyses(id: {eq: $vds_id}) {
+                    output
+                    sequencingGroups {
+                        id
+                    }
+                }
+            }
+        """,
+    )
+    query_results: dict[str, Any] = query(get_vds_analysis_query, variables={'vds_id': vds_id})
+    vds_path: str = query_results['analyses'][0]['output']
+    vds_sgids: list[str] = [sg['id'] for sg in query_results['analyses'][0]['sequencingGroups']]
+    return (vds_path, vds_sgids)
+
+
+def _process_withdrawals() -> None:
+    pass
+
+
+def combiner(cohort: Cohort, output_vds_path: str, save_path: str) -> PythonJob:
+    workflow_config = config_retrieve('workflow')
+    combiner_config = config_retrieve('combiner')
+
+    tmp_prefix = slugify(
+        f'{cohort.analysis_dataset.tmp_prefix}/{workflow_config["cohort"]}-{workflow_config["sequencing_type"]}-{combiner_config["vds_version"]}',
+    )
+
+    vds_paths: list[str] = []
+    sg_ids_in_vds: list[str] = []
+    new_sg_gvcfs: list[str] = []
+    cohort_sgs: list[SequencingGroup] = cohort.get_sequencing_groups(only_active=True)
+
+    if combiner_config.get('vds_analysis_ids', None) is not None:
+        for vds_id in combiner_config['vds_analysis_ids']:
+            tmp_query_res, tmp_sg_ids_in_vds = get_vds_ids_output(vds_id)
+            vds_paths.append(tmp_query_res)
+            sg_ids_in_vds = sg_ids_in_vds + tmp_sg_ids_in_vds
+
+    if combiner_config.get('merge_only_vds', False) is not True:
+        # Get SG IDs from the cohort object itself, rather than call Metamist.
+        # Get VDS IDs first and filter out from this list
+        new_sg_gvcfs = [str(sg.gvcf) for sg in cohort_sgs if sg.gvcf is not None and sg.id not in sg_ids_in_vds]
+
+    if new_sg_gvcfs and len(new_sg_gvcfs) == 0 and len(vds_paths) <= 1:
+        raise Exception
+
+    sequencing_group_names: list[str] = [
+        str(sg.id) for sg in cohort_sgs if sg.gvcf is not None and sg.id not in sg_ids_in_vds
+    ]
+
+    combiner_job: PythonJob = _initalise_combiner_job(cohort=cohort)
+
+    # Default to GRCh38 for reference if not specified
+    combiner_job.call(
+        _run,
+        output_vds_path=output_vds_path,
+        sequencing_type=workflow_config['sequencing_type'],
+        tmp_prefix=tmp_prefix,
+        genome_build=genome_build(),
+        save_path=save_path,
+        force_new_combiner=config_retrieve(['combiner', 'force_new_combiner']),
+        sequencing_group_names=sequencing_group_names,
+        gvcf_external_header=new_sg_gvcfs[0],
+        gvcf_paths=new_sg_gvcfs,
+        vds_paths=vds_paths,
+    )
+
+    return combiner_job
+
+
+def _run(
     output_vds_path: str,
     sequencing_type: str,
     tmp_prefix: str,
