+++ cpg_workflows/stages/large_cohort.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING, Any, Final, Tuple
 from cpg_utils import Path
 from cpg_utils.config import config_retrieve, genome_build, get_config, image_path
 from cpg_utils.hail_batch import get_batch, query_command
+from cpg_workflows.large_cohort.combiner import combiner
 from cpg_workflows.targets import Cohort, SequencingGroup
 from cpg_workflows.utils import slugify
 from cpg_workflows.workflow import (
@@ -13,13 +14,9 @@ from cpg_workflows.workflow import (
     get_workflow,
     stage,
 )
-from metamist.graphql import gql, query
 
 if TYPE_CHECKING:
-    from graphql import DocumentNode
-
     from hailtop.batch.job import PythonJob
-
 
 HAIL_QUERY: Final = 'hail query'
 
@@ -31,7 +28,7 @@ class Combiner(CohortStage):
         workflow_config = config_retrieve('workflow')
         combiner_config = config_retrieve('combiner')
         output_vds_name: str = slugify(
-            f"{cohort.name}-{workflow_config['sequencing_type']}-{combiner_config['vds_version']}",
+            f'{cohort.name}-{workflow_config["sequencing_type"]}-{combiner_config["vds_version"]}',
         )
 
         # include the list of all VDS IDs in the plan name
@@ -45,80 +42,13 @@ class Combiner(CohortStage):
             'combiner_plan': str(self.get_stage_cohort_prefix(cohort, 'tmp') / f'{combiner_plan_name}.json'),
         }
 
-    def get_vds_ids_output(self, vds_id: int) -> Tuple[str, list[str]]:
-        get_vds_analysis_query: DocumentNode = gql(
-            """
-            query getVDSByAnalysisIds($vds_id: Int!) {
-                analyses(id: {eq: $vds_id}) {
-                    output
-                    sequencingGroups {
-                        id
-                    }
-                }
-            }
-        """,
-        )
-        query_results: dict[str, Any] = query(get_vds_analysis_query, variables={'vds_id': vds_id})
-        vds_path: str = query_results['analyses'][0]['output']
-        vds_sgids: list[str] = [sg['id'] for sg in query_results['analyses'][0]['sequencingGroups']]
-        return (vds_path, vds_sgids)
-
     def queue_jobs(self, cohort: Cohort, inputs: StageInput) -> StageOutput | None:
-        # Can't import it before all configs are set:
-        from cpg_workflows.large_cohort import combiner
-
-        workflow_config = config_retrieve('workflow')
-        combiner_config = config_retrieve('combiner')
-
         output_paths = self.expected_outputs(cohort)
-        tmp_prefix = slugify(
-            f"{self.tmp_prefix}/{workflow_config['cohort']}-{workflow_config['sequencing_type']}-{combiner_config['vds_version']}",
-        )
 
-        # create these as empty lists instead of None, they have the same truthiness
-        vds_paths: list[str] = []
-        sg_ids_in_vds: list[str] = []
-        new_sg_gvcfs: list[str] = []
-
-        if combiner_config.get('vds_analysis_ids', None) is not None:
-            for vds_id in combiner_config['vds_analysis_ids']:
-                tmp_query_res, tmp_sg_ids_in_vds = self.get_vds_ids_output(vds_id)
-                vds_paths.append(tmp_query_res)
-                sg_ids_in_vds = sg_ids_in_vds + tmp_sg_ids_in_vds
-
-        if combiner_config.get('merge_only_vds', False) is not True:
-            # Get SG IDs from the cohort object itself, rather than call Metamist.
-            # Get VDS IDs first and filter out from this list
-            cohort_sgs: list[SequencingGroup] = cohort.get_sequencing_groups(only_active=True)
-            new_sg_gvcfs = [str(sg.gvcf) for sg in cohort_sgs if sg.gvcf is not None and sg.id not in sg_ids_in_vds]
-
-        if new_sg_gvcfs and len(new_sg_gvcfs) == 0 and len(vds_paths) <= 1:
-            return self.make_outputs(cohort, output_paths)
-
-        j: PythonJob = get_batch().new_python_job('Combiner', (self.get_job_attrs() or {}) | {'tool': HAIL_QUERY})
-        j.image(config_retrieve(['workflow', 'driver_image']))
-        j.memory(combiner_config.get('memory'))
-        j.storage(combiner_config.get('storage'))
-
-        # set this job to be non-spot (i.e. non-preemptible)
-        # previous issues with preemptible VMs led to multiple simultaneous QOB groups processing the same data
-        j.spot(config_retrieve(['combiner', 'preemptible_vms'], False))
-
-        # Default to GRCh38 for reference if not specified
-        j.call(
-            combiner.run,
+        j: PythonJob = combiner(
+            cohort=cohort,
             output_vds_path=str(output_paths['vds']),
-            sequencing_type=workflow_config['sequencing_type'],
-            tmp_prefix=tmp_prefix,
-            genome_build=genome_build(),
             save_path=output_paths['combiner_plan'],
-            force_new_combiner=config_retrieve(['combiner', 'force_new_combiner']),
-            sequencing_group_names=[
-                str(sg.id) for sg in cohort_sgs if sg.gvcf is not None and sg.id not in sg_ids_in_vds
-            ],
-            gvcf_external_header=new_sg_gvcfs[0],
-            gvcf_paths=new_sg_gvcfs,
-            vds_paths=vds_paths,
         )
 
         return self.make_outputs(cohort, output_paths, [j])
@@ -462,7 +392,7 @@ class Vqsr(CohortStage):
 
         # sed command to swap Float SB to Integer in-place and allow any length
         reheader_job.command(
-            fr"sed -i 's/<ID=SB,Number=1,Type=Float/<ID=SB,Number=.,Type=Integer/' {reheader_job.ofile}",
+            rf"sed -i 's/<ID=SB,Number=1,Type=Float/<ID=SB,Number=.,Type=Integer/' {reheader_job.ofile}",
         )
 
         b.write_output(
